Protein -> SimpleProtein
