+++ ms2/src/org/labkey/ms2/protein/ProteinServiceImpl.java
@@ -25,10 +25,12 @@ import org.labkey.api.cache.CacheManager;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.protein.PeptideCharacteristic;
 import org.labkey.api.protein.ProteinFeature;
+import org.labkey.api.protein.ProteinManager;
 import org.labkey.api.protein.ProteinPlus;
 import org.labkey.api.protein.ProteinSchema;
 import org.labkey.api.protein.ProteinService;
 import org.labkey.api.protein.Replicate;
+import org.labkey.api.protein.SimpleProtein;
 import org.labkey.api.protein.fasta.FastaDbLoader;
 import org.labkey.api.protein.fasta.FastaProtein;
 import org.labkey.api.protein.organism.GuessOrgByParsing;
@@ -64,6 +66,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.function.Consumer;
 import java.util.regex.Pattern;
 
 public class ProteinServiceImpl implements ProteinService
@@ -166,38 +169,39 @@ public class ProteinServiceImpl implements ProteinService
         return Collections.unmodifiableList(_peptideSearchViewProviders);
     }
 
-    @Override
-    public WebPartView<?> getProteinCoverageView(int seqId, List<PeptideCharacteristic> peptideCharacteristics, int aaRowWidth, boolean showEntireFragmentInCoverage, @Nullable String accessionForFeatures)
+    private WebPartView<?> getProteinCoverageView(int seqId, List<PeptideCharacteristic> peptideCharacteristics, int aaRowWidth, boolean showEntireFragmentInCoverage, @Nullable String accessionForFeatures, Consumer<ProteinViewBean> beanModifier)
     {
         ProteinViewBean bean = new ProteinViewBean();
-        bean.protein = ProteinManager.getProtein(seqId);
+        bean.protein = org.labkey.ms2.protein.ProteinManager.getProtein(seqId);
         bean.protein.setShowEntireFragmentInCoverage(showEntireFragmentInCoverage);
         bean.protein.setCombinedPeptideCharacteristics(peptideCharacteristics);
         bean.features = getProteinFeatures(accessionForFeatures);
         bean.aaRowWidth = aaRowWidth;
+        beanModifier.accept(bean);
         return new JspView<>("/org/labkey/ms2/protein/view/proteinCoverageMap.jsp", bean);
+    }
+
+    @Override
+    public WebPartView<?> getProteinCoverageView(int seqId, List<PeptideCharacteristic> peptideCharacteristics, int aaRowWidth, boolean showEntireFragmentInCoverage, @Nullable String accessionForFeatures)
+    {
+        return getProteinCoverageView(seqId, peptideCharacteristics, aaRowWidth, showEntireFragmentInCoverage, accessionForFeatures, bean -> {});
     }
 
     @Override
     public WebPartView<?> getProteinCoverageViewWithSettings(int seqId, List<PeptideCharacteristic> peptideCharacteristics, int aaRowWidth, boolean showEntireFragmentInCoverage, @Nullable String accessionForFeatures , List<Replicate> replicates, List<PeptideCharacteristic> modifiedPeptideCharacteristics, boolean showStackedPeptides)
     {
-        ProteinViewBean bean = new ProteinViewBean();
-        bean.protein = ProteinManager.getProtein(seqId);
-        bean.protein.setShowEntireFragmentInCoverage(showEntireFragmentInCoverage);
-        bean.protein.setCombinedPeptideCharacteristics(peptideCharacteristics);
-        bean.features = getProteinFeatures(accessionForFeatures);
-        bean.aaRowWidth = aaRowWidth;
-        bean.replicates = replicates;
-        bean.showViewSettings = true;
-        bean.protein.setModifiedPeptideCharacteristics(modifiedPeptideCharacteristics);
-        bean.protein.setShowStakedPeptides(showStackedPeptides);
-        return new JspView<>("/org/labkey/ms2/protein/view/proteinCoverageMap.jsp", bean);
+        return getProteinCoverageView(seqId, peptideCharacteristics, aaRowWidth, showEntireFragmentInCoverage, accessionForFeatures, bean -> {
+            bean.replicates = replicates;
+            bean.showViewSettings = true;
+            bean.protein.setModifiedPeptideCharacteristics(modifiedPeptideCharacteristics);
+            bean.protein.setShowStakedPeptides(showStackedPeptides);
+        });
     }
 
     @Override
     public WebPartView<?> getAnnotationsView(int seqId, Map<String, Collection<HtmlString>> extraAnnotations)
     {
-        Protein protein = ProteinManager.getProtein(seqId);
+        SimpleProtein protein = ProteinManager.getProtein(seqId);
         return new AnnotationView(protein, extraAnnotations);
     }
